split domain, fiber, sheaf classes
add fiber.index and fiber.coords methods
