+++ topos/base/__init__.py
@@ -1,9 +1,6 @@
 from .nerve         import Nerve
 from .complex       import Complex
 from .graph         import Graph
-from .domain        import Domain, Fiber, Sheaf
-# old
-#from .set           import Set, Iterable, Map, Mapping
-#from .fiber         import Fiber
-#from .simplex       import Sequence, Chain, Face, Simplex, Seq
-#from .hypergraph    import Poset, Hypergraph
+from .sheaf         import Sheaf
+from .fiber         import Fiber
+from .domain        import Domain
